+++ src/components/EducationSection/PreviousEntries.jsx
@@ -1,27 +1,26 @@
 import { ActionIcon, Button } from "@mantine/core";
-import { IconPencil, IconTrashX } from "@tabler/icons-react";
-import DateSelector from "./DateSelector";
+import { IconPencil, IconTrashX, IconCirclePlus } from "@tabler/icons-react";
 
-export default function EducationSection() {
-  const previouslyAddedEntries = [
-    {
-      degree: "Masters of Data Science",
-      major: "",
-      school: "Univeristy of Adelaide",
-      startDate: "September 2020",
-      endDate: "Present",
-    },
-    {
-      degree: "Bachelor of Commerce",
-      major: "Finance/Accounting",
-      school: "Univeristy of New South wales",
-      startDate: "March 2014",
-      endDate: "November 2017",
-    },
-  ];
+export default function PreviousEntries({ handleAddNewEntry }) {
+  // const previouslyAddedEntries = [
+  //   {
+  //     degree: "Masters of Data Science",
+  //     major: "",
+  //     school: "Univeristy of Adelaide",
+  //     startDate: "September 2020",
+  //     endDate: "Present",
+  //   },
+  //   {
+  //     degree: "Bachelor of Commerce",
+  //     major: "Finance/Accounting",
+  //     school: "Univeristy of New South wales",
+  //     startDate: "March 2014",
+  //     endDate: "November 2017",
+  //   },
+  // ];
 
   return (
-    <article className="educationSection">
+    <>
       <ul className="previousEntries">
         <li className="previousEntry">
           <div className="summary">
@@ -33,7 +32,7 @@ export default function EducationSection() {
             <ActionIcon
               variant="transparent"
               color="gray"
-              size="xl"
+              size="lg"
               aria-label="Edit"
             >
               <IconPencil style={{ width: "70%", height: "70%" }} />
@@ -41,7 +40,7 @@ export default function EducationSection() {
             <ActionIcon
               variant="transparent"
               color="gray"
-              size="xl"
+              size="lg"
               aria-label="Delete"
             >
               <IconTrashX style={{ width: "70%", height: "70%" }} />
@@ -61,7 +60,7 @@ export default function EducationSection() {
             <ActionIcon
               variant="transparent"
               color="gray"
-              size="xl"
+              size="lg"
               aria-label="Edit"
             >
               <IconPencil style={{ width: "70%", height: "70%" }} />
@@ -69,7 +68,7 @@ export default function EducationSection() {
             <ActionIcon
               variant="transparent"
               color="gray"
-              size="xl"
+              size="lg"
               aria-label="Delete"
             >
               <IconTrashX style={{ width: "70%", height: "70%" }} />
@@ -77,44 +76,17 @@ export default function EducationSection() {
           </div>
         </li>
       </ul>
-      <ul>
-        <div className="grid">
-          <li>
-            <label htmlFor="degree">Degree</label>
-            <input type="text" id="degree" placeholder="Bachelor of Commerce" />
-          </li>
-          <li>
-            <label htmlFor="major">Major</label>
-            <input type="text" id="major" placeholder="Finance/Accounting" />
-          </li>
-        </div>
-        <li>
-          <label htmlFor="school">School</label>
-          <input
-            type="text"
-            id="school"
-            placeholder="University of New South Wales"
-          />
-        </li>
-        <li className="dateSelectors">
-          <ul className="grid">
-            <DateSelector />
-          </ul>
-        </li>
-      </ul>
 
-      <div className="button-group">
-        <Button variant="outline" color="#00ccb8" className="primary-button">
-          Save
-        </Button>
-        <Button
-          variant="outline"
-          color="#8490b3"
-          className="destructive-button"
-        >
-          Cancel
-        </Button>
-      </div>
-    </article>
+      <Button
+        leftSection={<IconCirclePlus />}
+        color="#00ccb8"
+        variant="outline"
+        size="md"
+        className="addAnotherButton"
+        onClick={handleAddNewEntry}
+      >
+        Add another education
+      </Button>
+    </>
   );
 }
